feat: Add GET support, deprecate `requestData`
The client could only perform POST requests. This PR adds support for
GET requests. The method used is made explicit in the api, the
`requestData` method is deprecated. `postData` is a drop-in replacement
for this function, and has the same method signature. The `requestData`
method will be removed in the next major release.

To perform a GET, use the `getData` method. It is used the same way the
POST method is, but has no `body` parameter.